Reorganize Tests (#985)
We no longer need a dedicated folder for Python3.6+ tests
We no longer need to check six.PY3 in tests
