+++ graphene/types/tests/test_enum.py
@@ -1,5 +1,3 @@
-import six
-
 from ..argument import Argument
 from ..enum import Enum, PyEnum
 from ..field import Field
@@ -115,9 +113,6 @@ def test_enum_from_builtin_enum_accepts_lambda_description():
 
 
 def test_enum_from_python3_enum_uses_enum_doc():
-    if not six.PY3:
-        return
-
     from enum import Enum as PyEnum
 
     class Color(PyEnum):
