+++ graphene/types/tests/test_objecttype.py
@@ -265,3 +265,17 @@ def test_abstract_objecttype_can_str():
         field = MyScalar()
 
     assert str(MyObjectType) == "MyObjectType"
+
+
+def test_objecttype_meta_with_annotations():
+    class Query(ObjectType):
+        class Meta:
+            name: str = "oops"
+
+        hello = String()
+
+        def resolve_hello(self, info):
+            return "Hello"
+
+    schema = Schema(query=Query)
+    assert schema is not None
