+++ tox.ini
@@ -9,7 +9,7 @@ deps =
 setenv =
      PYTHONPATH = .:{envdir}
 commands =
-    py{36,37}: py.test --cov=graphene graphene examples tests_asyncio tests_py36 {posargs}
+    py{36,37}: py.test --cov=graphene graphene examples tests_asyncio {posargs}
 
 [testenv:pre-commit]
 basepython=python3.6
